pyproject.toml
[tool.poetry]
name = "runpod-serverless"
version = "0.1.0"
description = ""
authors = ["AshDavid12 <gaashdavid@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.1"
pytube = "^15.0.0"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
fastapi = "^0.111.1"
requests = "^2.32.3"
httpx = "^0.27.0"
uvicorn = "^0.30.1"
asyncio = "^3.4.3"
confluent-kafka = "^2.5.0"
flask = "^3.0.3"
faster-whisper = "1.0.3"
runpod = "^1.7.0"
librosa = ">=0.10.2.post1,<0.11.0"
soundfile = "^0.12.1"
openai = "^1.42.0"
numpy = "^1.22.0"
torch = "2.1.0"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
